Update README.md
Although Exceptions are the most obvious solution (for people using languages where exceptions are a norm), no one's said that this will actually be it. "Better error handling features" could be a lot of things, e.g. a standardized way of passing result/error as an enum and a built-in operator to chain operations based on that. Could be. We don't know.